PR fixes, increate minor version
