+++ semantics/src/main/scala/aqua/semantics/header/Picker.scala
@@ -120,68 +120,4 @@ object Picker {
           )
   }
 
-  given [S[_]]: Picker[LspContext[S]] with {
-
-    private def ops[S[_]](ctx: LspContext[S]) = PickerOps[RawContext](ctx.raw)
-
-    override def blank: LspContext[S] = LspContext[S](Picker[RawContext].blank, Map.empty)
-    override def exports(ctx: LspContext[S]): Option[Map[String, Option[String]]] = ops(ctx).exports
-    override def funcNames(ctx: LspContext[S]): List[String] = ops(ctx).funcNames
-
-    override def addPart(ctx: LspContext[S], part: (LspContext[S], RawPart)): LspContext[S] =
-      ctx.copy(raw = ops(ctx).addPart(part._1.raw -> part._2))
-
-    override def setInit(ctx: LspContext[S], ctxInit: Option[LspContext[S]]): LspContext[S] =
-      ctx.copy(raw = ops(ctx).setInit(ctxInit.map(_.raw)))
-
-    override def all(ctx: LspContext[S]): Set[String] =
-      ops(ctx).all
-    override def module(ctx: LspContext[S]): Option[String] = ops(ctx).module
-    override def declares(ctx: LspContext[S]): Set[String] = ops(ctx).declares
-
-    override def setAbility(ctx: LspContext[S], name: String, ctxAb: LspContext[S]): LspContext[S] =
-      ctx.copy(raw = ops(ctx).setAbility(name, ctxAb.raw))
-
-    override def setModule(
-      ctx: LspContext[S],
-      name: Option[String],
-      declares: Set[String]
-    ): LspContext[S] =
-      ctx.copy(raw = ops(ctx).setOptModule(name, declares))
-
-    override def setExports(
-      ctx: LspContext[S],
-      exports: Map[String, Option[String]]
-    ): LspContext[S] =
-      ctx.copy(raw = ops(ctx).setExports(exports))
-
-    override def pick(
-      ctx: LspContext[S],
-      name: String,
-      rename: Option[String],
-      declared: Boolean
-    ): Option[LspContext[S]] =
-      ops(ctx)
-        .pick(name, rename, declared)
-        .map(rc =>
-          ctx.copy(
-            raw = rc,
-            abDefinitions =
-              ctx.abDefinitions.get(name).fold(Map.empty)(t => Map(rename.getOrElse(name) -> t)),
-            rootArrows =
-              ctx.rootArrows.get(name).fold(Map.empty)(t => Map(rename.getOrElse(name) -> t)),
-            constants =
-              ctx.constants.get(name).fold(Map.empty)(t => Map(rename.getOrElse(name) -> t))
-          )
-        )
-
-    override def pickHeader(ctx: LspContext[S]): LspContext[S] =
-      ctx.copy(raw = ops(ctx).pickHeader)
-
-    override def pickDeclared(
-      ctx: LspContext[S]
-    )(implicit semi: Semigroup[LspContext[S]]): LspContext[S] =
-      ctx.copy(raw = ops(ctx).pickDeclared)
-  }
-
 }
